Moved `FoundationDataBuilder.get` to `../network`.
